We modified the code to use the ColorMatch API
With this modification we can now read all the sample colors reliably with or without lexan glass and within the 1-10 centimeter range. We have only tested this in flourescent light and we still don't know if the color is a fraction away from changing so we still need to do more testing. - Rushil, Shaun