Left joystick button now does nothing if intake roller is stopped.
The previous commit would have meant that the intake roller was turned on whenever the left joystick button was pressed, regardless of whether or not it had been turned on before. Now it only causes the direction to reverse if the intake roller is already enabled. The button press is ignored if the roller is stopped.